+++ docs/release-notes.md
@@ -8,6 +8,10 @@ Under development.
 * Raised minimum MediaWiki version from 1.31 to 1.35
 * Raised minimum PHP version from 7.1 to 7.4.3
 
+### Chameleon 3.4.0
+
+Under development.
+
 ### Chameleon 3.3.0
 
 Released on June 20, 2021.